requirements.txt
# Minimal, pinned versions for reliable builds
runpod==1.6.2
torch==2.2.0
diffusers==0.30.3
transformers==4.44.2
safetensors==0.4.5
sentencepiece==0.2.0
Pillow==10.4.0
numpy==1.24.4
requests==2.32.3
protobuf==4.25.4
accelerate==0.33.0

# LoRA training dependencies
peft==0.12.0
datasets==2.20.0
bitsandbytes==0.43.1
scipy==1.11.4
tqdm==4.66.4